refactor: 리네임, 함수 분리, 에러메시지 수정
